--- backend/src/main/java/com/musinsa/backend/brand/mapper/BrandMapper.java
@@ -1,4 +1,0 @@
-package com.musinsa.backend.brand.mapper;
-
-public class BrandMapper {
-}
