--- backend/src/main/java/com/musinsa/backend/brand/model/Brand.java
@@ -1,4 +1,0 @@
-package com.musinsa.backend.brand.model;
-
-public class Brand {
-}
