--- backend/src/main/java/com/musinsa/backend/brand/service/BrandServiceImpl.java
@@ -1,4 +1,0 @@
-package com.musinsa.backend.brand.service;
-
-public class BrandServiceImpl {
-}
