+++ backend/src/main/java/com/musinsa/backend/internal/mapper/BrandMapper.java
@@ -1,0 +1,4 @@
+package com.musinsa.backend.internal.mapper;
+
+public class BrandMapper {
+}
