+++ backend/src/main/java/com/musinsa/backend/internal/model/Brand.java
@@ -1,0 +1,4 @@
+package com.musinsa.backend.internal.model;
+
+public class Brand {
+}
