+++ backend/src/main/java/com/musinsa/backend/internal/service/BrandService.java
@@ -1,0 +1,4 @@
+package com.musinsa.backend.internal.service;
+
+public interface BrandService {
+}
