+++ backend/src/main/java/com/musinsa/backend/internal/service/BrandServiceImpl.java
@@ -1,0 +1,4 @@
+package com.musinsa.backend.internal.service;
+
+public class BrandServiceImpl {
+}
