--- backend/src/main/java/com/musinsa/backend/search/controller/SearchController.java
@@ -1,8 +1,0 @@
-package com.musinsa.backend.search.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class SearchController {
-
-}
